backend/requirements.txt
flask
flask-cors
azure-cosmos
python-dotenv
gunicorn
PyJWT
werkzeug
bcrypt
openai
azure-search-documents
azure-storage-blob==12.19.0
